Update/reorganize tests
