+++ test/functional/dynamic/test_delete.py
@@ -1,7 +1,9 @@
-import pytest
-from pytest_bdd import scenario, then, parsers
+# coding=utf-8
+"""delete feature tests.
+All step implementations are in conftest.py
+"""
 
-from openshift.dynamic import exceptions
+from pytest_bdd import scenario
 
 
 @scenario('delete.feature', 'Delete a resource that exists')
@@ -14,11 +16,3 @@ def test_delete_a_resource_that_exists():
 def test_delete_a_resource_that_does_not_exist():
     """Delete a resource that does not exist"""
     pass
-
-
-@then(parsers.parse('<group_version>.<kind> <name> does not exist in <namespace>'))
-def resource_not_in_namespace(admin_client, group_version, kind, name, namespace):
-    """<group_version>.<kind> <name> does not exist in <namespace>."""
-    resource = admin_client.resources.get(api_version=group_version, kind=kind)
-    with pytest.raises(exceptions.NotFoundError):
-        resource.get(name, namespace)
